+++ tests/files/Modules/Bitrix/BitrixDetailPageFilesTest.php
@@ -17,7 +17,9 @@ class BitrixDetailPageFilesTest extends BitrixTestCase{
 	function tearDown(){
 		parent::tearDown();
 
-		$this->module->deleteFolder();
+		if ($this->module){
+			$this->module->deleteFolder();
+		}
 	}
 
 	/** @test */
