+++ tests/interface/Modules/Bitrix/BitrixDetailPageInterfaceTest.php
@@ -17,7 +17,9 @@ class BitrixDetailPageInterfaceTest extends BitrixTestCase{
 	function tearDown(){
 		parent::tearDown();
 
-		$this->module->deleteFolder();
+		if ($this->module){
+			$this->module->deleteFolder();
+		}
 	}
 
 	/** @test */
