+++ tests/interface/Modules/Bitrix/BitrixEventHandlersInterfaceTest.php
@@ -20,7 +20,9 @@ class BitrixEventHandlersInterfaceTest extends BitrixTestCase{
 	function tearDown(){
 		parent::tearDown();
 
-		$this->module->deleteFolder();
+		if ($this->module){
+			$this->module->deleteFolder();
+		}
 	}
 
 	/** @test */
